new vision encoder impl

+++ moondream/hf/vision_encoder.py
@@ -242,7 +242,7 @@ class VisionEncoder(nn.Module):
             [
                 Resize(size=im_size, interpolation=InterpolationMode.BICUBIC),
                 ToImage(),
-                ToDtype(torch.float32, scale=True),
+                ToDtype(torch.float16, scale=True),
                 Normalize(mean=[0.5, 0.5, 0.5], std=[0.5, 0.5, 0.5]),
             ]
         )(image)
